Cambios en los perfiles de los torneos

+++ domino/domino/services/event.py
@@ -199,17 +199,17 @@ def delete(request: Request, event_id: str, db: Session):
             db_event.updated_date = datetime.now()
             db.commit()
             
-            # if db_event.image:
-            user_created = get_one_by_username(db_event.created_by, db=db)
-            path = "/public/events/" + str(user_created.id) + "/" + str(db_event.id) + "/"
-            try:
-                del_image(path=path, name=str(db_event.image))
-            except:
-                pass
-            try:
-                remove_dir(path=path[:-1])
-            except:
-                pass
+            if db_event.image:
+                user_created = get_one_by_username(db_event.created_by, db=db)
+                path = "/public/events/" + str(user_created.id) + "/" + str(db_event.id) + "/"
+                try:
+                    del_image(path=path, name=str(db_event.image))
+                except:
+                    pass
+                try:
+                    remove_dir(path=path[:-1])
+                except:
+                    pass
                 
             return result
         else:
